Update xCAT dracut module for Fedora36.
This move away from using a bare-minimum dracut-built image to using a
more complete systemd-based image, with doxcat run from the pre-mount
hook rather than the cmdline hook. This is an attempt to leverage the
modern kernel module loading infrastructure to fix issues with failure
to load network and storage drivers.