Update doxcat for the new genesis environment.
Since we're explicitly setting up a number of things via dracut we can
skip several elements of the system setup which are handled in doxcat at
the moment. Currently the main changes are in the rsyslog and chrony
setup, but we're also checking for a few other things that are set up by
dracut - dhcp networking, and rpcbind and friends. We should be able to
move more out of doxcat into dracut modules/hooks, however this is still
under consideration.

This also includes whitespace fixes, as the file had a mix of tabs and
spaces along with confused indentation - the diff is considerably larger
than it would be otherwise, but the inconsistent indentation made it
difficult to break out the whitepsace fixes from everything else.