Ensure additional nics are brought up.
This needs to happen even if the bootnic has come up without needing
additional intervention.

Also, don't try to bring up nics that have NO-CARRIER - this is
pointless, unless we want to assume that something underneath us will
change the state of the hardware. Given this a discovery/utility image
that doesn't seem like a great idea.