Rework BMC IP address determination.
The old way doesn't seem to work reliably at all - on some hardware
`ipmitool sol info 2` works, on others it needs to be 1. There's also no
apparent relationship with the payload channel listed. So instead, this
new logic just tries channels until it finds one that works, and uses
that.