Fix setup of root account.
There was a conflict between the way that the base dracut module set up
root and the way that we modified that in order to support our use case.
This change resolves that conflict, leaving us with a single entry for
root in /etc/passwd rather than two different entries.