Update user handling in xCAT dracut module.
We were ending up with a completely messed up set of user accounts
because a number of things that required users in the final system were
moved over to using existing modules, which of course handled the users
themselves.

We also moved to a smarter model of managing the users we do want to set
up, copying them across from the source system rather than trying to set
them up from scratch - this will give us a more consistent set of
userids.