Changed the way that restart decides what to restart.
This was triggered by the realisation that in some situations the
dhclient pid files, and even the lease files, can end up going away - in
theory this shouldn't happen, because the lease file at least will be
left behind until it's cleared up either by dhclient or something else,
but in my testing there was no certainty about that. Instead, we now use
the `dynamic` flag from the `ip a s` output to select interfaces.

I've also moved the lldpad setup down to the end of the network bring-up
section, since it makes more sense to do that after things have finished
being set up.