Set hostname to something more useful early on.
If DHCP doesn't give us a hostname (which is normal during discovery) we
end up with hostname `(none)`, which isn't very informative. Instead, if
our destiny is discover, then set the hostname to
`discover-<bootif-mac>` nice and early, so that we can ensure that
things like rsyslogd pick it up.

In theory this has potential to confuse other parts of dracut, but as
far as I can tell hostname='(none)' is only checked once, and that's in
order to avoid using a completely useless value - nothing else seems to
care.